fix double -> single conversion

+++ systems/robotInterfaces/@Biped/planSwingPitched.m
@@ -129,7 +129,7 @@ end
 % terrain_pts_in_world = T_toe_local_to_world \ terrain_pts_in_toe_local;
 % max_terrain_ht_in_world = max(terrain_pts_in_world(3,:))
 if ~isempty(swing2.terrain_pts)
-  max_terrain_ht_in_world = max(swing2.terrain_pts(2,:));
+  max_terrain_ht_in_world = double(max(swing2.terrain_pts(2,:)));
 else
   max_terrain_ht_in_world = -inf;
 end
